Using single classes for live/standard metrics to avoid duplication SpanProcessor used in LiveMetricExporter instead of other way around Updating example
