+++ azure_monitor/src/azure_monitor/sdk/auto_collection/__init__.py
@@ -15,9 +15,11 @@ from azure_monitor.sdk.auto_collection.performance_metrics import (
     PerformanceMetrics,
 )
 from azure_monitor.sdk.auto_collection.request_metrics import RequestMetrics
+from azure_monitor.sdk.auto_collection.utils import AutoCollectionType
 
 __all__ = [
     "AutoCollection",
+    "AutoCollectionType",
     "AzureMetricsSpanProcessor",
     "DependencyMetrics",
     "RequestMetrics",
@@ -40,8 +42,11 @@ class AutoCollection:
         labels: Dict[str, str],
         span_processor: AzureMetricsSpanProcessor,
     ):
-        self._performance_metrics = PerformanceMetrics(meter, labels)
+        col_type = AutoCollectionType.STANDARD_METRICS
+        self._performance_metrics = PerformanceMetrics(meter, labels, col_type)
         self._dependency_metrics = DependencyMetrics(
-            meter, labels, span_processor
+            meter, labels, span_processor, col_type
         )
-        self._request_metrics = RequestMetrics(meter, labels, span_processor)
+        self._request_metrics = RequestMetrics(
+            meter, labels, span_processor, col_type
+        )
