+++ azure_monitor/src/azure_monitor/sdk/auto_collection/live_metrics/__init__.py
@@ -5,29 +5,25 @@ from typing import Dict
 
 from opentelemetry.metrics import Meter
 
-from azure_monitor.sdk.auto_collection.live_metrics.dependency_metrics import (
-    DependencyLiveMetrics,
+from azure_monitor.sdk.auto_collection import AutoCollectionType
+from azure_monitor.sdk.auto_collection.dependency_metrics import (
+    DependencyMetrics,
 )
 from azure_monitor.sdk.auto_collection.live_metrics.exporter import (
     LiveMetricsExporter,
 )
-from azure_monitor.sdk.auto_collection.live_metrics.performance_metrics import (
-    PerformanceLiveMetrics,
-)
-from azure_monitor.sdk.auto_collection.live_metrics.request_metrics import (
-    RequestLiveMetrics,
+from azure_monitor.sdk.auto_collection.live_metrics.manager import (
+    LiveMetricsManager,
 )
 from azure_monitor.sdk.auto_collection.metrics_span_processor import (
     AzureMetricsSpanProcessor,
 )
+from azure_monitor.sdk.auto_collection.performance_metrics import (
+    PerformanceMetrics,
+)
+from azure_monitor.sdk.auto_collection.request_metrics import RequestMetrics
 
-__all__ = [
-    "LiveMetricsAutoCollection",
-    "LiveMetricsExporter",
-    "DependencyLiveMetrics",
-    "PerformanceLiveMetrics",
-    "RequestLiveMetrics",
-]
+__all__ = ["LiveMetricsAutoCollection"]
 
 
 class LiveMetricsAutoCollection:
@@ -44,11 +40,16 @@ class LiveMetricsAutoCollection:
         meter: Meter,
         labels: Dict[str, str],
         span_processor: AzureMetricsSpanProcessor,
+        instrumentation_key: str,
     ):
-        self._performance_metrics = PerformanceLiveMetrics(meter, labels)
-        self._dependency_metrics = DependencyLiveMetrics(
-            meter, labels, span_processor
+        col_type = AutoCollectionType.LIVE_METRICS
+        self._performance_metrics = PerformanceMetrics(meter, labels, col_type)
+        self._dependency_metrics = DependencyMetrics(
+            meter, labels, span_processor, col_type
         )
-        self._request_metrics = RequestLiveMetrics(
-            meter, labels, span_processor
+        self._request_metrics = RequestMetrics(
+            meter, labels, span_processor, col_type
+        )
+        self._manager = LiveMetricsManager(
+            meter, instrumentation_key, span_processor
         )
