+++ azure_monitor/src/azure_monitor/sdk/auto_collection/utils.py
@@ -1,0 +1,11 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+from enum import Enum
+
+
+class AutoCollectionType(Enum):
+    """Automatic collection of metrics type
+    """
+
+    STANDARD_METRICS = 0
+    LIVE_METRICS = 1
