+++ azure_monitor/tests/auto_collection/live_metrics/test_manager.py
@@ -16,6 +16,9 @@ from azure_monitor.sdk.auto_collection.live_metrics.manager import (
     LiveMetricsPing,
     LiveMetricsPost,
 )
+from azure_monitor.sdk.auto_collection.metrics_span_processor import (
+    AzureMetricsSpanProcessor,
+)
 
 
 # pylint: disable=protected-access
@@ -33,6 +36,7 @@ class TestLiveMetricsManager(unittest.TestCase):
         cls._manager = None
         cls._ping = None
         cls._post = None
+        cls._span_processor = AzureMetricsSpanProcessor()
 
     @classmethod
     def tearDownClass(cls):
@@ -55,6 +59,7 @@ class TestLiveMetricsManager(unittest.TestCase):
             self._manager = LiveMetricsManager(
                 meter=self._meter,
                 instrumentation_key=self._instrumentation_key,
+                span_processor=self._span_processor,
             )
             self.assertFalse(self._manager._is_user_subscribed)
             self.assertEqual(
@@ -72,6 +77,7 @@ class TestLiveMetricsManager(unittest.TestCase):
             self._manager = LiveMetricsManager(
                 meter=self._meter,
                 instrumentation_key=self._instrumentation_key,
+                span_processor=self._span_processor,
             )
             self._manager.interval = 60
             time.sleep(1)
@@ -127,7 +133,10 @@ class TestLiveMetricsManager(unittest.TestCase):
                 200, None, {"x-ms-qps-subscribed": "false"}
             )
             self._post = LiveMetricsPost(
-                exporter=LiveMetricsExporter(self._instrumentation_key),
+                exporter=LiveMetricsExporter(
+                    self._instrumentation_key,
+                    span_processor=self._span_processor,
+                ),
                 meter=self._meter,
                 instrumentation_key=self._instrumentation_key,
             )
@@ -144,7 +153,10 @@ class TestLiveMetricsManager(unittest.TestCase):
                 200, None, {"x-ms-qps-subscribed": "true"}
             )
             self._post = LiveMetricsPost(
-                exporter=LiveMetricsExporter(self._instrumentation_key),
+                exporter=LiveMetricsExporter(
+                    self._instrumentation_key,
+                    span_processor=self._span_processor,
+                ),
                 meter=self._meter,
                 instrumentation_key=self._instrumentation_key,
             )
@@ -156,7 +168,10 @@ class TestLiveMetricsManager(unittest.TestCase):
         with mock.patch("requests.post") as request:
             request.return_value = MockResponse(400, None, {})
             self._post = LiveMetricsPost(
-                exporter=LiveMetricsExporter(self._instrumentation_key),
+                exporter=LiveMetricsExporter(
+                    self._instrumentation_key,
+                    span_processor=self._span_processor,
+                ),
                 meter=self._meter,
                 instrumentation_key=self._instrumentation_key,
             )
