+++ azure_monitor/tests/auto_collection/test_request_metrics.py
@@ -11,6 +11,7 @@ from azure_monitor.sdk.auto_collection import request_metrics
 from azure_monitor.sdk.auto_collection.metrics_span_processor import (
     AzureMetricsSpanProcessor,
 )
+from azure_monitor.sdk.auto_collection.utils import AutoCollectionType
 
 
 # pylint: disable=protected-access
@@ -29,21 +30,26 @@ class TestRequestMetrics(unittest.TestCase):
     def setUp(self):
         request_metrics.requests_map.clear()
 
-    def test_constructor(self):
+    def test_constructor_standard_metrics(self):
         mock_meter = mock.Mock()
         request_metrics_collector = request_metrics.RequestMetrics(
             meter=mock_meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         self.assertEqual(request_metrics_collector._meter, mock_meter)
         self.assertEqual(request_metrics_collector._labels, self._test_labels)
-
-        self.assertEqual(mock_meter.register_observer.call_count, 2)
-
+        self.assertEqual(mock_meter.register_observer.call_count, 3)
         create_metric_calls = mock_meter.register_observer.call_args_list
-
         create_metric_calls[0].assert_called_with(
+            callback=request_metrics_collector._track_request_failed_rate,
+            name="\\ApplicationInsights\\Requests Failed/Sec",
+            description="Incoming Requests Failed Rate",
+            unit="rps",
+            value_type=int,
+        )
+        create_metric_calls[1].assert_called_with(
             callback=request_metrics_collector._track_request_duration,
             name="\\ASP.NET Applications(??APP_W3SVC_PROC??)\\Request Execution Time",
             description="Incoming Requests Average Execution Time",
@@ -51,10 +57,30 @@ class TestRequestMetrics(unittest.TestCase):
             value_type=int,
         )
 
-        create_metric_calls[1].assert_called_with(
+        create_metric_calls[2].assert_called_with(
             callback=request_metrics_collector._track_request_rate,
             name="\\ASP.NET Applications(??APP_W3SVC_PROC??)\\Requests/Sec",
-            description="Incoming Requests Average Execution Rate",
+            description="Incoming Requests Rate",
+            unit="rps",
+            value_type=int,
+        )
+
+    def test_constructor_live_metrics(self):
+        mock_meter = mock.Mock()
+        request_metrics_collector = request_metrics.RequestMetrics(
+            meter=mock_meter,
+            labels=self._test_labels,
+            span_processor=self._span_processor,
+            collection_type=AutoCollectionType.LIVE_METRICS,
+        )
+        self.assertEqual(request_metrics_collector._meter, mock_meter)
+        self.assertEqual(request_metrics_collector._labels, self._test_labels)
+        self.assertEqual(mock_meter.register_observer.call_count, 1)
+        create_metric_calls = mock_meter.register_observer.call_args_list
+        create_metric_calls[0].assert_called_with(
+            callback=request_metrics_collector._track_request_failed_rate,
+            name="\\ApplicationInsights\\Requests Failed/Sec",
+            description="Incoming Requests Failed Rate",
             unit="rps",
             value_type=int,
         )
@@ -64,6 +90,7 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         self._span_processor.request_duration = 0.1
         self._span_processor.request_count = 10
@@ -86,6 +113,7 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         self._span_processor.request_duration = 0.1
         self._span_processor.request_count = 10
@@ -109,6 +137,7 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         time_mock.time.return_value = 100
         request_metrics.requests_map["last_time"] = 98
@@ -133,6 +162,7 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         request_metrics.requests_map["last_time"] = None
         obs = Observer(
@@ -154,6 +184,7 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         time_mock.time.return_value = 100
         request_metrics.requests_map["last_rate"] = 5
@@ -167,6 +198,77 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
         )
         request_metrics_collector._track_request_rate(obs)
+        self.assertEqual(
+            obs.aggregators[tuple(self._test_labels.items())].current, 5
+        )
+
+    @mock.patch("azure_monitor.sdk.auto_collection.request_metrics.time")
+    def test_track_request_failed_rate(self, time_mock):
+        request_metrics_collector = request_metrics.RequestMetrics(
+            meter=self._meter,
+            labels=self._test_labels,
+            span_processor=self._span_processor,
+            collection_type=AutoCollectionType.LIVE_METRICS,
+        )
+        time_mock.time.return_value = 100
+        request_metrics.requests_map["last_time"] = 98
+        self._span_processor.failed_request_count = 4
+        obs = Observer(
+            callback=request_metrics_collector._track_request_failed_rate,
+            name="test",
+            description="test",
+            unit="test",
+            value_type=int,
+            meter=self._meter,
+        )
+        request_metrics_collector._track_request_failed_rate(obs)
+        self.assertEqual(
+            obs.aggregators[tuple(self._test_labels.items())].current, 2
+        )
+
+    @mock.patch("azure_monitor.sdk.auto_collection.request_metrics.time")
+    def test_track_request_failed_rate_time_none(self, time_mock):
+        time_mock.time.return_value = 100
+        request_metrics_collector = request_metrics.RequestMetrics(
+            meter=self._meter,
+            labels=self._test_labels,
+            span_processor=self._span_processor,
+            collection_type=AutoCollectionType.LIVE_METRICS,
+        )
+        request_metrics.requests_map["last_time"] = None
+        obs = Observer(
+            callback=request_metrics_collector._track_request_failed_rate,
+            name="test",
+            description="test",
+            unit="test",
+            value_type=int,
+            meter=self._meter,
+        )
+        request_metrics_collector._track_request_failed_rate(obs)
+        self.assertEqual(
+            obs.aggregators[tuple(self._test_labels.items())].current, 0
+        )
+
+    @mock.patch("azure_monitor.sdk.auto_collection.request_metrics.time")
+    def test_track_request_failed_rate_error(self, time_mock):
+        request_metrics_collector = request_metrics.RequestMetrics(
+            meter=self._meter,
+            labels=self._test_labels,
+            span_processor=self._span_processor,
+            collection_type=AutoCollectionType.LIVE_METRICS,
+        )
+        time_mock.time.return_value = 100
+        request_metrics.requests_map["last_rate"] = 5
+        request_metrics.requests_map["last_time"] = 100
+        obs = Observer(
+            callback=request_metrics_collector._track_request_failed_rate,
+            name="test",
+            description="test",
+            unit="test",
+            value_type=int,
+            meter=self._meter,
+        )
+        request_metrics_collector._track_request_failed_rate(obs)
         self.assertEqual(
             obs.aggregators[tuple(self._test_labels.items())].current, 5
         )
